requirements.txt
aiofiles
fastapi
msrest
oauth2client
openpyxl
pandas
Pillow
pyodbc
requests
schedule
starlette
tinys3
uvicorn
xlrd
azure-storage-blob
azure-cognitiveservices-vision-computervision
python-multipart
azure-storage-file-share
uvloop
httptools
azure-storage-queue
xlsxwriter
opencv-python
nltk
ftputil
iptcinfo3
python-xmp-toolkit
piexif
exifread
python-multipart